rename query hooks
