+++ client/src/components/NewTag/use-new-tag.ts
@@ -1,5 +1,5 @@
+import useTagsQuery from "@/lib/query/useTagsQuery";
 import { queryClient } from "@lib/query-client";
-import useTagsQuery from "@lib/query/use-tags-query";
 import { useTagSelection } from "@lib/state/selected-tags-state";
 import useAuthentication from "@lib/use-authentication";
 import type { NewTag } from "@type/server/tag.types";
