+++ client/src/components/Register/use-register.ts
@@ -1,4 +1,4 @@
-import useRegisterMutation from "@/lib/query/use-register-mutation";
+import useRegisterMutation from "@/lib/query/useRegisterMutation";
 import type { NewUser } from "@type/server/user.types";
 import { useState } from "react";
