+++ client/src/components/Today/Notes.tsx
@@ -1,7 +1,7 @@
 import { filterTagsById } from "@/lib/filter-tags";
+import useNotesQuery from "@/lib/query/useNotesQuery";
+import useTagsQuery from "@/lib/query/useTagsQuery";
 import { isToday } from "@lib/datetime/compare";
-import useNotesQuery from "@lib/query/use-notes-query";
-import useTagsQuery from "@lib/query/use-tags-query";
 import { Note } from "./Note";
 import S from "./style/Today.style";
