+++ client/src/lib/query/useActivitiesQuery.ts
