+++ client/src/lib/use-authentication.ts
@@ -1,6 +1,6 @@
-import useLoginMutation from "./query/use-login-mutation";
-import useLogoutMutation from "./query/use-logout-mutation";
-import useMeQuery from "./query/use-me-query";
+import useLoginMutation from "./query/useLoginMutation";
+import useLogoutMutation from "./query/useLogoutMutation";
+import useMeQuery from "./query/useMeQuery";
 
 export default function useAuthentication() {
 	const { mutate: login } = useLoginMutation();
